+++ client/src/helper/cDefault.ts
@@ -1,0 +1,5 @@
+export const cDefault: string = `
+class Solution:
+  def main():
+    # Write code here
+`;
