+++ client/src/helper/cppDefault.ts
@@ -1,0 +1,5 @@
+export const cppDefault: string = `
+class Solution:
+  def main():
+    # Write code here
+`;
