+++ client/src/helper/pythonDefault.ts
@@ -1,0 +1,5 @@
+export const pythonDefault: string = `
+class Solution:
+  def main(input):
+    # Write code here, do not rename main
+`;
