+++ client/src/helper/pythonDriver.ts
@@ -1,0 +1,9 @@
+export const pythonDriver : string = `
+  output = []
+  ls
+  for i in ls:
+  # function should return output of format specified
+    output.append(main(i))
+  for o in output:
+    print(o)
+`;
